Bugfix/performance enhancement method argument checking
The first argument in a method must be an instance of the correct type.
This patch fixes a bug (check result was always true) and adds compiler
code to speedup checking. The classes get a __md5__ property, that is
defined on compile time and can be used to identify these.

The duplicate function pyjs___isinstance is replaced by pyjslib._isinstance

Unittest added to test first argument checking

git-svn-id: https://pyjamas.svn.sourceforge.net/svnroot/pyjamas/trunk@611 7a2bd370-bda8-463c-979e-2900ccfb811e