Simplified import for pyjs
* Modules are stored in $pyjs.modules
 * Modules are also accessible via $pyjs.modules_hash (to prevent some
   pyjs_eval statements)
 * The global module_load_request is removed

Some comment on module loading:

All to-be-imported module names are in app_modlist of class Modload.
Since we're only _loading_ the modules there, we can do that in almost
any order. There's one limitation: a child/sub module cannot be loaded
unless its parent is loaded. It has to be chained in the module list.
  (1) $pyjs.modules.pyjamas
  (2) $pyjs.modules.pyjamas.ui
  (3) $pyjs.modules.pyjamas.ui.Widget
Therefore, all modules are collected and sorted on depth (i.e. the
number of dots in it)
As long as we don't move on to the next depth unless all modules of the
previous depth are loaded, we won't run into unchainable modules
The execution of the module code is done when the import statement is
reached, or after loading all the modules for the main module.

If the modules are static, the main module is started right away.

git-svn-id: https://pyjamas.svn.sourceforge.net/svnroot/pyjamas/trunk@683 7a2bd370-bda8-463c-979e-2900ccfb811e