Removed manipulation of body overflow attribute in showGlass/hideGlass methods calling Window.enableScrolling()
Changing this attribute causes layouts to reflow, which can
produce unnecessary re-rendering delay and visual changes,
despite the fact that using GlassWidget is supposed to be
undetectable.

Removing the manipulation of this attribute prevents reflow
caused by showing/hiding the GlassWidget, although it does
raise the possibility of the mouse moving off of the widget
if the browser window is scrolled during capture operations.

It is felt that this is a small and unlikely problem compared
to the disruption caused by changing the overflow attribute.